chore(fork): Change package from "atombender" to "arikkfir"
This chore is needed so that it will be possible to create an interim
release until PRs to upstream "github.com/atombender/go-jsonschema" are
accepted.

It is the goal to avoid using releases from repository in favor of the
upstream repository whenever possible.

This change also adjusts GitHub Actions workflows to accommodate for the
change (e.g. secrets, Homebrew repository, etc.)